requirements.txt
python-telegram-bot[job-queue]
feedparser
beautifulsoup4
requests
transformers
torch
python-dotenv
huggingface_hub[hf_xet]